refactor(plotting): move GSPlot and GSPlotParameter imports to single_plot module
