+++ src/goesdl/plotting/__init__.py
@@ -1,3 +1,3 @@
-from ..plotting.preview import GSPlot, GSPlotParameter
+from .single_plot import GSPlot, GSPlotParameter
 
 __all__ = ["GSPlot", "GSPlotParameter"]
